Relocate index.ts to constants folder
This commit moves the index.ts file to the constants folder and updates the imports in the test files accordingly.
